requirements.txt
Django==4.2.0
djangorestframework==3.14.0
pytest==7.3.1
django-cors-headers==4.3.0
